App Fixes (#74)
* Refactor Project Name & Packages

Originally had project name set as "EmptyActivity", now it is "GymPlanner" to be seamless with our actual name chosen. Same was done for all packages & in google_services.json

* Ran AGP Upgrade Assistant

To resolve Issue #68

* Removed Exercise List Screen

Exercise list screen didn't fit into our vision of the final state of our project, so it is now removed.

* SHA-1 Certificate gotten

* What iff i blew my brains out

* Navbar Accessibility improvements

* Exercise Input Accessibility Improvements

* Added excessive Documentation

* Overall Code Improvements + Addition of documentation

Updated to use a List instead of concatenating a string + UI overhaul of displaying the list. Documentation added.

* Code Cleanup